Set schedule size according to oldest file
by dynamically adjusting the size of the schedule until it is large
enough to encompass the oldest back (for exp, fib and simple counting).
This fixes #6.

For gauss I did not fully understand the code, so I did not change it.

Added a regression test to the “test suite” in the manpage.